update station characterization
